sort out the funciton to make the species names list in order from the tree
