+++ src/make_CAFE_input_from_orthofinder.py
@@ -1,19 +1,11 @@
 import pandas as pd
 import re
 from collections import Counter
+import src.parse_gff as gff
 
 # add a functional description column to the Orthogroups.GeneCount.tsv output to be able to run CAFE
 # also make sure that the species tree names match the columns in the file
 
-def split_at_second_occurrence(s, char = "_"): # split the gene string at the second occurence of "_" to get only the species name
-    if s.count(char)<2:
-        return s
-    else:
-        first_occurrence = s.find(char)
-        second_occurrence = s.find(char, first_occurrence + 1)
-        species = s[:second_occurrence]
-        return species
-    
 
 def read_orthogroups_file(path_to_orthogroups_file):
     orthogroups_dict = {} # dictionary containing {orthogroup_ID : [list, of, orthogroup, member, gene_IDs]}
@@ -22,19 +14,6 @@ def read_orthogroups_file(path_to_orthogroups_file):
             orthogroup = orthogroup.strip().split(":")
             orthogroups_dict[orthogroup[0]] = orthogroup[1].strip().split(" ")
     return orthogroups_dict
-
-
-def make_species_order_from_tree(newick_tree_path):
-    # Regular expression to extract leaf names
-    # This matches strings between commas, parentheses, and before colons.
-    leaf_pattern = r'(?<=\(|,)([a-zA-Z0-9_]+)(?=:)'
-    with open(newick_tree_path, "r") as newick_tree_file:
-        newick_tree_string = newick_tree_file.readlines()[0]
-        # print(newick_tree_string)
-        leaf_names = re.findall(leaf_pattern, newick_tree_string)
-        # leaf names are like "A_obtectus_filtered_proteinfasta" but we only care about the species names in the beginning
-        species_names = [split_at_second_occurrence(leaf, "_") for leaf in leaf_names]
-    return species_names
 
 
 def get_tree_name_from_header(header_str, species_names):
@@ -77,7 +56,7 @@ def parse_table_from_txt(orthogroups_filepath, outfile_name, species_names = "")
     table_rows = []
     for orthogroup_ID, transcripts_list in orthogroups_dict.items():
         # count the number of times each species appears
-        transcripts_list = [split_at_second_occurrence(transcript) for transcript in transcripts_list]        
+        transcripts_list = [gff.split_at_second_occurrence(transcript) for transcript in transcripts_list]        
         transcripts_dict = Counter(transcripts_list)
         transcripts_dict = {species : int(num) for species, num in transcripts_dict.items()}
 
@@ -143,7 +122,7 @@ if __name__ == '__main__':
 
 
     # species_names = make_species_order_from_tree("/Users/miltr339/Box Sync/code/annotation_pipeline/annotation_scripts_ordered/14_species_orthofinder_tree.nw")
-    species_names = make_species_order_from_tree("/Users/milena/Box Sync/code/annotation_pipeline/annotation_scripts_ordered/14_species_orthofinder_tree.nw")
+    species_names = gff.make_species_order_from_tree("/Users/milena/Box Sync/code/annotation_pipeline/annotation_scripts_ordered/14_species_orthofinder_tree.nw")
 
     orthoDB_orthogroups = "/Users/miltr339/work/orthofinder/orthoDB_orthogroups/Orthogroups.GeneCount.tsv"
     orthoDB_uniform_masked_orthogroups = "/Users/milena/work/orthofinder/orthoDB_uniform_masking_orthogroups/N0.tsv"
@@ -158,5 +137,5 @@ if __name__ == '__main__':
 
     # parse_table_from_txt(orthoDB_filtered_orthogroups, "CAFE_input_orthoDB_filtered.tsv", species_names)
     # parse_table_from_txt(orthoDB_filtered_orthogroups, "CAFE_input_orthoDB_filtered.tsv")
-    species_names = make_species_order_from_tree("/Users/milena/work/orthofinder/orthoDB_TE_filtered_orthogroups/SpeciesTree_rooted.txt")
+    species_names = gff.make_species_order_from_tree("/Users/milena/work/orthofinder/orthoDB_TE_filtered_orthogroups/SpeciesTree_rooted.txt")
     modify_orthogroups_from_N0(orthoDB_TE_filtered_orthogroups, species_names=species_names)
