+++ src/single_exon_stats.py
@@ -1,7 +1,7 @@
 ### calculate gff statistics focused on single- and multi-exon genes
 ## write the results to stdout
 
-import parse_gff
+import src.parse_gff as gff
 import sys
 
 try:
@@ -16,7 +16,7 @@ try:
         print("Invalid third argument. Please pass 'True' or 'False'.")
         sys.exit(1)
 
-    parse_gff.print_single_exon_stats(orthodb_annot, include_list=boolean_value) # include or exclude a list of all the single-exon IDs
+    gff.print_single_exon_stats(orthodb_annot, include_list=boolean_value) # include or exclude a list of all the single-exon IDs
     # write to stdout, so if you include the list a long list of gene IDs will be written to stdout
     # parse_gff.print_single_exon_stats(native_annot, include_list=boolean_value)
