--- tests/test_dummy.py
@@ -1,5 +1,0 @@
-import unittest
-
-class TestSanity(unittest.TestCase):
-    def test_true(self):
-        self.assertTrue(True)
